write endpoint functions in controllers file
